refactor: split 5 files over 500-line limit
- threads_view.rs (680) → threads_view/mod.rs (358) + messages.rs (343)
- actions/mod.rs (660) → mod.rs (451) + threads.rs (283)
  - Merged navigation.rs into helpers.rs to free file slot
- lifecycle.rs (570) → lifecycle.rs (359) + run/threads.rs (222)
  - Thread methods as free fns (avoids #[expect] for multiple_inherent_impl)
- llm_types.rs (538) → llm_types.rs (121) + config/models.rs (420)
  - Model enums in sibling module, no pub use re-export (banned by clippy)
  - Fixed 3 import sites: actions.rs, runtime.rs, llms/mod.rs
- overview/mod.rs (508) → mod.rs (475) + visualizers.rs (38)

All files now under 500 lines. All callbacks green.
