+++ src/llms/mod.rs
@@ -26,8 +26,10 @@ use crate::state::Message;
 
 // Re-export LLM types from cp-base so that `crate::llms::LlmProvider` etc. work
 pub(crate) use cp_base::config::llm_types::{
-    AnthropicModel, ApiCheckResult, ClaudeCodeV2Model, DeepSeekModel, GrokModel, GroqModel, LlmProvider, MiniMaxModel,
-    ModelInfo, StreamEvent,
+    ApiCheckResult, LlmProvider, ModelInfo, StreamEvent,
+};
+pub(crate) use cp_base::config::models::{
+    AnthropicModel, ClaudeCodeV2Model, DeepSeekModel, GrokModel, GroqModel, MiniMaxModel,
 };
 
 // Re-export provider clients through the module path for get_client()
